+++ v2/container/container_close_value_test.go
@@ -1,0 +1,51 @@
+package container
+
+import (
+    "sync"
+    "testing"
+
+    containercontract "github.com/precision-soft/melody/v2/container/contract"
+)
+
+type valueCloser struct {
+    counter *int
+    lock    *sync.Mutex
+}
+
+func (instance valueCloser) Close() error {
+    instance.lock.Lock()
+    defer instance.lock.Unlock()
+
+    *instance.counter++
+
+    return nil
+}
+
+func TestContainer_Close_ValueTypeServiceClosedOnce(t *testing.T) {
+    serviceContainer := NewContainer()
+
+    var lock sync.Mutex
+    count := 0
+
+    MustRegister[valueCloser](
+        serviceContainer,
+        "value.closer",
+        func(resolver containercontract.Resolver) (valueCloser, error) {
+            return valueCloser{counter: &count, lock: &lock}, nil
+        },
+    )
+
+    _ = MustFromResolver[valueCloser](serviceContainer, "value.closer")
+    _ = MustFromResolverByType[valueCloser](serviceContainer)
+
+    if err := serviceContainer.Close(); nil != err {
+        t.Fatalf("unexpected close error: %v", err)
+    }
+
+    lock.Lock()
+    defer lock.Unlock()
+
+    if 1 != count {
+        t.Fatalf("expected value-type service Close to be called once, got %d", count)
+    }
+}