+++ v2/validation/value.go
@@ -1,0 +1,35 @@
+package validation
+
+import (
+    "reflect"
+)
+
+func dereferenceValue(value any) (any, bool) {
+    if nil == value {
+        return nil, false
+    }
+
+    reflectedValue := reflect.ValueOf(value)
+
+    for {
+        kind := reflectedValue.Kind()
+
+        if (reflect.Pointer == kind) || (reflect.Interface == kind) {
+            if true == reflectedValue.IsNil() {
+                return nil, false
+            }
+
+            reflectedValue = reflectedValue.Elem()
+
+            continue
+        }
+
+        break
+    }
+
+    if reflect.Invalid == reflectedValue.Kind() {
+        return nil, false
+    }
+
+    return reflectedValue.Interface(), true
+}