+++ v3/container/container_close_test.go
@@ -190,6 +190,49 @@ func TestContainer_Close_ClosesDependentsBeforeDependencies_ByTypeResolution(t *
     }
 }
 
+type valueCloser struct {
+    counter *int
+    lock    *sync.Mutex
+}
+
+func (instance valueCloser) Close() error {
+    instance.lock.Lock()
+    defer instance.lock.Unlock()
+
+    *instance.counter++
+
+    return nil
+}
+
+func TestContainer_Close_ValueTypeServiceClosedOnce(t *testing.T) {
+    serviceContainer := NewContainer()
+
+    var lock sync.Mutex
+    count := 0
+
+    MustRegister[valueCloser](
+        serviceContainer,
+        "value.closer",
+        func(resolver containercontract.Resolver) (valueCloser, error) {
+            return valueCloser{counter: &count, lock: &lock}, nil
+        },
+    )
+
+    _ = MustFromResolver[valueCloser](serviceContainer, "value.closer")
+    _ = MustFromResolverByType[valueCloser](serviceContainer)
+
+    if err := serviceContainer.Close(); nil != err {
+        t.Fatalf("unexpected close error: %v", err)
+    }
+
+    lock.Lock()
+    defer lock.Unlock()
+
+    if 1 != count {
+        t.Fatalf("expected value-type service Close to be called once, got %d", count)
+    }
+}
+
 type circularServiceA struct{}
 type circularServiceB struct{}